Fix (composer-find) in aliased use statement.
When the cursor is on Foo\Bar in use Foo\Bar as Baz, it tries to expand
the name using the current namespace because of the lack of leading \.
Note that it works without the as part because we can resolve the name
by looking for matching use statement (which only matches on the alias,
if present) before resorting to expanding the name using the current
file's namespace.

This change fixes the issue by allowing matching a use statement with
alias when expanding fully-qualified name. Closes #1.