Prefer project root containing vendor directory
Previously when setting b:composer_root, the closest composer.json would
be used. Now, the closest composer.json with sibling vendor/autoload.php
is used. If none is found, falls back to the closest composer.json.

The rationale is that it is not useful to scope :Composer and the
navigation commands (:Ecomposer et al.) to an installed dependency, and
<Plug>(composer-find) was broken when jumping from a dependency.

Fixes #5.